insert_data/requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0 